feat: Update DataCreator project to generate fewer weather records
chore: Update .gitignore and remove unnecessary code

feat: Add Station and WeatherRecord models and DataCreator project

+++ src/Domain/StationStats.cs
@@ -3,8 +3,8 @@ public class StationStats
 {
     public string Name { get; set; }
     public int Count { get; set; }
-    public int Lowest { get; set; }
-    public int Highest { get; set; }
+    public double Lowest { get; set; }
+    public double Highest { get; set; }
     public double Average { get; set; }
-    public int Total { get; set; }
+    public double Total { get; set; }
 }
